[LLM-1784] - Organize spread out `utils.py` to be in `llm/_internal/common`
