+++ python/ray/llm/_internal/serve/configs/openai_api_models_patch.py
@@ -17,7 +17,7 @@ from pydantic import (
 )
 from typing_extensions import Annotated
 
-from ray.llm._internal.utils import try_import
+from ray.llm._internal.common.utils.import_utils import try_import
 
 if TYPE_CHECKING:
     from vllm.sampling_params import GuidedDecodingParams
