+++ python/ray/llm/_internal/serve/deployments/llm/image_retriever.py
@@ -5,7 +5,7 @@ from typing import TYPE_CHECKING
 
 import aiohttp
 
-from ray.llm._internal.utils import try_import
+from ray.llm._internal.common.utils.import_utils import try_import
 
 if TYPE_CHECKING:
     from PIL.Image import Image
