+++ python/ray/llm/_internal/serve/deployments/llm/vllm/vllm_engine.py
@@ -7,6 +7,7 @@ from concurrent.futures.thread import ThreadPoolExecutor
 from typing import TYPE_CHECKING, AsyncGenerator, List, Optional, Tuple
 
 import ray
+from ray.llm._internal.common.utils.import_utils import try_import
 from ray.llm._internal.serve.configs.constants import (
     MAX_NUM_TOPLOGPROBS_ALLOWED,
     MIN_NUM_TOPLOGPROBS_ALLOWED,
@@ -51,7 +52,6 @@ from ray.llm._internal.serve.observability.metrics.utils import (
     ClockUnit,
     MsClock,
 )
-from ray.llm._internal.utils import try_import
 from ray.util import metrics
 from ray.util.placement_group import PlacementGroup
 from ray.util.scheduling_strategies import PlacementGroupSchedulingStrategy
